Finish Webscrape till map is built.
After exploring the website, I've decided to focus more on the building of the map itself, then seeing how I want to populate it using the wvtourism website. The website layout is not as hierarchical as I thought, so even though I still want to scrape for the data, I'm not sure what all information I want just yet.