scraped things-to-do
Massive progress on scrapes. Going to use the actual listings instead of all companies as many of those look to be outdated, but I don't want to get rid of the files incase they could be used for something in the future. Now just need to finish scraping the other parts of the website, turn the addresses to lat/long, and start on the routes.